+++ RTRendererCUDAQT/RTRendererCUDAQT/Hittable.h
@@ -8,17 +8,17 @@ class Material;
 
 struct HitRecord
 {
-	double t;
+	float t;
 	Vec3 point;
 	Vec3 normal;
 	Material* matPtr;
-	double u;
-	double v;
+	float u;
+	float v;
 };
 
 class Hittable
 {
 public:
-	__device__ virtual bool hit(const Ray& r, double tMin, double tMax, HitRecord& rec, curandState* localRandState) const = 0;
+	__device__ virtual bool hit(const Ray& r, float tMin, float tMax, HitRecord& rec, curandState* localRandState) const = 0;
 	__device__ virtual bool boundingBox(AABB& box) const = 0;
 };
