+++ RTRendererCUDAQT/RTRendererCUDAQT/Vec3.h
@@ -5,7 +5,7 @@
 class Vec3
 {
 public:
-	double e[3];
+	float e[3];
 
 	__host__ __device__ Vec3() 
 	{
@@ -14,7 +14,7 @@ public:
 		e[2] = 0;
 	}
 
-	__host__ __device__ Vec3(double e0, double e1, double e2) 
+	__host__ __device__ Vec3(float e0, float e1, float e2) 
 	{
 		e[0] = e0;
 		e[1] = e1;
@@ -27,16 +27,16 @@ public:
 	__host__ __device__ inline Vec3& operator-=(const Vec3& v2);
 	__host__ __device__ inline Vec3& operator*=(const Vec3& v2);
 	__host__ __device__ inline Vec3& operator/=(const Vec3& v2);
-	__host__ __device__ inline Vec3& operator*=(const double v2);
-	__host__ __device__ inline Vec3& operator/=(const double v2);
+	__host__ __device__ inline Vec3& operator*=(const float v2);
+	__host__ __device__ inline Vec3& operator/=(const float v2);
 
 
-	__host__ __device__ inline double squaredLength() const { return e[0] * e[0] + e[1] * e[1] + e[2] * e[2]; }
-	__host__ __device__ inline double length() const { return sqrt(squaredLength()); }
+	__host__ __device__ inline float squaredLength() const { return e[0] * e[0] + e[1] * e[1] + e[2] * e[2]; }
+	__host__ __device__ inline float length() const { return sqrt(squaredLength()); }
 	__host__ __device__ inline void makeUnitVector();
 
-	__host__ __device__ inline void Vec3::writeFrameBuffer(int i, int j, double* fBuffer);
-	__host__ __device__ inline void Vec3::readFrameBuffer(int i, int j, double* fBuffer);
+	__host__ __device__ inline void Vec3::writeFrameBuffer(int i, int j, float* fBuffer);
+	__host__ __device__ inline void Vec3::readFrameBuffer(int i, int j, float* fBuffer);
 };
 
 inline std::istream& operator>>(std::istream& is, Vec3& t)
@@ -53,7 +53,7 @@ inline std::ostream& operator<<(std::ostream& os, Vec3& t)
 
 __host__ __device__ inline void Vec3::makeUnitVector()
 {
-	double k = 1.0 / squaredLength();
+	float k = 1.0 / squaredLength();
 	e[0] *= k;
 	e[1] *= k;
 	e[2] *= k;
@@ -95,7 +95,7 @@ __host__ __device__ inline Vec3 operator/(const Vec3& v1, const Vec3& v2)
 	);
 }
 
-__host__ __device__ inline Vec3 operator*(const Vec3& v1, double t)
+__host__ __device__ inline Vec3 operator*(const Vec3& v1, float t)
 {
 	return Vec3(
 		v1.e[0] * t,
@@ -104,12 +104,12 @@ __host__ __device__ inline Vec3 operator*(const Vec3& v1, double t)
 	);
 }
 
-__host__ __device__ inline Vec3 operator*(double t, const Vec3& v1)
+__host__ __device__ inline Vec3 operator*(float t, const Vec3& v1)
 {
 	return v1*t;
 }
 
-__host__ __device__ inline Vec3 operator/(const Vec3& v1, double t)
+__host__ __device__ inline Vec3 operator/(const Vec3& v1, float t)
 {
 	return Vec3(
 		v1.e[0] / t,
@@ -119,7 +119,7 @@ __host__ __device__ inline Vec3 operator/(const Vec3& v1, double t)
 }
 
 
-__host__ __device__ inline double dot(const Vec3& v1, const Vec3& v2)
+__host__ __device__ inline float dot(const Vec3& v1, const Vec3& v2)
 {
 	return v1.e[0] * v2.e[0] +
 		   v1.e[1] * v2.e[1] +
@@ -168,7 +168,7 @@ __host__ __device__ inline Vec3& Vec3::operator/=(const Vec3& v2)
 	return *this;
 }
 
-__host__ __device__ inline Vec3& Vec3::operator*=(const double t)
+__host__ __device__ inline Vec3& Vec3::operator*=(const float t)
 {
 	e[0] *= t;
 	e[1] *= t;
@@ -176,7 +176,7 @@ __host__ __device__ inline Vec3& Vec3::operator*=(const double t)
 	return *this;
 }
 
-__host__ __device__ inline Vec3& Vec3::operator/=(const double t)
+__host__ __device__ inline Vec3& Vec3::operator/=(const float t)
 {
 	e[0] /= t;
 	e[1] /= t;
@@ -189,7 +189,7 @@ __host__ __device__ inline Vec3 unitVector(Vec3 v)
 	return v / v.length();
 }
 
-__host__ __device__ inline void Vec3::writeFrameBuffer(int i, int j, double* fBuffer)
+__host__ __device__ inline void Vec3::writeFrameBuffer(int i, int j, float* fBuffer)
 {
 	int index = (MAX_Y - j - 1) * MAX_X * 3 + i * 3; // Find location in frame buffer.
 
@@ -204,7 +204,7 @@ __host__ __device__ inline void Vec3::writeFrameBuffer(int i, int j, double* fBu
 #endif
 }
 
-__host__ __device__ inline void Vec3::readFrameBuffer(int i, int j, double* fBuffer)
+__host__ __device__ inline void Vec3::readFrameBuffer(int i, int j, float* fBuffer)
 {
 	int index = (MAX_Y - j - 1) * MAX_X * 3 + i * 3; // Find location in frame buffer.
